almost

+++ app/Http/Controllers/ArticleController.php
@@ -6,6 +6,11 @@ use App\Models\Article;
 use App\Models\User;
 use App\Models\Comment;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
+use App\Events\NewArticleEvent;
+use Illuminate\Support\Facades\Cache;
+use Illuminate\Support\Facades\DB;
+
 
 class ArticleController extends Controller
 {
@@ -14,7 +19,10 @@ class ArticleController extends Controller
      */
     public function index()
     {
-        $articles = Article::latest()->paginate(6);
+        $page = isset($_GET['page']) ? $_GET['page'] : 0;
+        $articles = Cache::remember('articles'.$page, 3000, function(){
+            return Article::latest()->paginate(6);
+        });
         return view('article.index', ['articles'=> $articles]);
     }
 
@@ -31,6 +39,11 @@ class ArticleController extends Controller
      */
     public function store(Request $request)
     {
+        $keys = DB::table('cache')->whereRaw('`key` GLOB :key', [':key'=>'articles*[0-9]'])->get();
+        foreach($keys as $param){
+            Cache::forget($param->key);
+        }
+        Gate::authorize('create', [self::class]);
         $request->validate([
             'date'=>'date',
             'name'=>'required|min:5|max:100',
@@ -41,8 +54,10 @@ class ArticleController extends Controller
         $article->name = $request->name;
         $article->desc = $request->desc;
         $article->user_id = 1;
-        $article->save();
-        return redirect('/article');
+        if ($article->save()) {
+            NewArticleEvent::dispatch($article);
+            return redirect('/article');
+        }
     }
 
     /**
@@ -50,9 +65,19 @@ class ArticleController extends Controller
      */
     public function show(Article $article)
     {
-        $comments = Comment::where('article_id', $article->id)->get();
-        $user = User::findOrFail($article->user_id);
-        return view('article.show', ['article'=>$article, 'user'=>$user, 'comments'=>$comments]);
+        if (isset($_GET['notify'])) auth()->user()->notifications->where('id', $_GET['notify'])->first()->markAsRead();
+        $result = Cache::rememberForever('comment_article'.$article->id, function()use($article){
+            $comments = Comment::where('article_id', $article->id)
+                            ->where('accept', true)
+                            ->get();
+            $user = User::findOrFail($article->user_id);
+            return [
+                'comments'=>$comments,
+                'user'=>$user
+            ];
+        });
+        return view('article.show', ['article'=>$article, 'user'=>$result['user'], 'comments'=>$result['comments']]);
+
     }
 
     /**
@@ -69,6 +94,11 @@ class ArticleController extends Controller
      */
     public function update(Request $request, Article $article)
     {
+        $keys = DB::table('cache')->whereRaw('`key` GLOB :key', [':key'=>'articles*[0-9]'])->get();
+        foreach($keys as $param){
+            Cache::forget($param->key);
+        }
+        Gate::authorize('update', $article);
         $request->validate([
             'date'=>'date',
             'name'=>'required|min:5|max:100',
@@ -87,6 +117,8 @@ class ArticleController extends Controller
      */
     public function destroy(Article $article)
     {
+        Cache::flush();
+        Gate::authorize('delete', $article);
         if ($article->delete()) return redirect('/article')->with('status','Delete success');
         else return redirect()->route('article.show', ['article'=>$article->id])->with('status','Delete don`t success');
     }
